fix warnings

+++ assembly/nmd_x86_emulator.c
@@ -79,7 +79,7 @@ bool nmd_x86_emulate(NMD_X86Cpu* cpu, size_t maxCount)
 		NMD_X86Instruction instruction;
 		const void* buffer = NMD_GET_PHYSICAL_ADDRESS(cpu->rip);
 		const bool validBuffer = NMD_IN_BOUNDARIES(buffer);
-		if (!validBuffer || !nmd_x86_decode_buffer(buffer, endVirtualAddress - cpu->rip, &instruction, (NMD_X86_MODE)cpu->mode, NMD_X86_DECODER_FLAGS_MINIMAL))
+		if (!validBuffer || !nmd_x86_decode_buffer(buffer, (size_t)(endVirtualAddress - cpu->rip), &instruction, (NMD_X86_MODE)cpu->mode, NMD_X86_DECODER_FLAGS_MINIMAL))
 		{
 			if (cpu->callback)
 				cpu->callback(cpu, &instruction, validBuffer ? NMD_X86_EMULATOR_EXCEPTION_BAD_INSTRUCTION : NMD_X86_EMULATOR_EXCEPTION_BAD_MEMORY);
